components reorganization
