+++ src/components/WelcomePage.js
@@ -1,0 +1,10 @@
+import {WindowDiv} from './styles/styles';
+
+export default () => {
+    return(
+        <WindowDiv>
+            <header>This is my welcome message</header>
+            <button>Enter</button>
+        </WindowDiv>
+    )
+}
