+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App/App.js';
+import App from './components/App.js';
 
 ReactDOM.render(
   <App />,
